Move UR_DeveloperSettings
