+++ Source/OpenTournament/GameModes/UR_BotCreationComponent.cpp
@@ -8,7 +8,7 @@
 #include "GameFramework/PlayerState.h"
 #include "Kismet/GameplayStatics.h"
 
-#include "UR_DeveloperSettings.h"
+#include "Development/UR_DeveloperSettings.h"
 #include "UR_GameMode.h"
 #include "GameModes/UR_ExperienceManagerComponent.h"
